% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crop_map}
\alias{crop_map}
\title{Crops a map to a new bounding box}
\usage{
crop_map(map, new_bb)
}
\description{
This function reduces the dimensions of a map to fit a given bounding box
as closely as possible.  The bounding box is readjusted to fit pixel
boundaries.
}

