% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_bb}
\alias{expand_bb}
\title{Compute a query set to obtain tiles that cover a given bounding box}
\usage{
expand_bb(bb, new_bb, stepwidth = 0.9)
}
\arguments{
\item{bb,new_bb}{Data frames, each with one row and four columns
\code{\{ll|ur\}.\{lat|lon\}},
representing the base and the desired bounding boxes}

\item{stepwidth}{Relative width of each step, defaults to \code{0.9}.}
}
\description{
This function assumes that \code{bb} is the bounding box that results from
a certain call to \code{\link[ggmap]{get_map}}.  From this, this function
computes a set of queries (given as latitude/longitude pairs) that will
result in maps covering the target bounding box when passed to otherwise
identical \code{get_map} calls.
}

