% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ggmapStitch}
\alias{ggmapStitch}
\alias{ggmapStitch-package}
\title{ggmapStitch.}
\description{
The ggmap package offers downloading map tiles from a variety of
providers, but (at least for Google Maps) the maps are restricted to a maximum
size.  This package removes this restriction by providing a function to
construct maps for given bounding box and zoom level.
}

