% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stitch_map}
\alias{stitch_map}
\title{Downloads a map with a specified bounding box, stitches tiles if necessary}
\usage{
stitch_map(bb, stepwidth = 1, ...)
}
\arguments{
\item{bb}{A data frame with one row and four columns \code{\{ll|ur\}.\{lat|lon\}}}

\item{stepwidth}{The default will attempt to download tiles that perfectly align.
This may fail due to rounding errors.  Use a value slightly less than 1 to allow
for an overlap in the downloaded tiles.}

\item{...}{Further arguments passed to \code{\link[ggmap]{get_map}}}
}
\description{
This creates a map that corresponds to a given bounding box by repeatedly
calling \code{\link[ggmap]{get_map}} and seamlessly stitching the resulting
tiles.
}

